monve InterceptorHelper code to MongoUtils
