+++ plugins/mongodb/src/main/java/com/megaease/easeagent/plugin/mongodb/interceptor/MongoInternalConnectionSendAndReceiveAsync4MetricInterceptor.java
@@ -29,6 +29,7 @@ import com.megaease.easeagent.plugin.enums.Order;
 import com.megaease.easeagent.plugin.interceptor.MethodInfo;
 import com.megaease.easeagent.plugin.interceptor.NonReentrantInterceptor;
 import com.megaease.easeagent.plugin.mongodb.MongoPlugin;
+import com.megaease.easeagent.plugin.mongodb.MongoUtils;
 import com.megaease.easeagent.plugin.mongodb.points.MongoDBInternalConnectionPoints;
 import com.mongodb.event.CommandEvent;
 import com.mongodb.event.CommandFailedEvent;
@@ -70,12 +71,12 @@ public class MongoInternalConnectionSendAndReceiveAsync4MetricInterceptor implem
 //            LOGGER.info("SingleResultCallbackProxy onResult metric");
             this.delegate.onResult(result, t);
             Context context = EaseAgent.getContext();
-            CommandEvent event = context.get(InterceptorHelper.EVENT_KEY);
+            CommandEvent event = context.get(MongoUtils.EVENT_KEY);
             if (event == null) {
                 return;
             }
-            MongoMetric mongoMetric = this.mongoCtx.get(InterceptorHelper.METRIC);
-            AutoRefreshPluginConfigImpl config = this.mongoCtx.get(InterceptorHelper.CONFIG);
+            MongoMetric mongoMetric = this.mongoCtx.get(MongoUtils.METRIC);
+            AutoRefreshPluginConfigImpl config = this.mongoCtx.get(MongoUtils.CONFIG);
             if (event instanceof CommandSucceededEvent) {
                 MetricHelper.commandSucceeded(context, config, mongoMetric, (CommandSucceededEvent) event);
             } else if (event instanceof CommandFailedEvent) {
